AetherLearn/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
motor>=3.4.0
beanie>=1.23.0
python-dotenv==1.0.0
google-cloud-aiplatform>=1.38.0  # Latest Vertex AI SDK
google-generativeai>=0.3.0  # Google Generative AI library
vertexai>=0.5.0  # Updated Vertex AI Python SDK
google-api-python-client>=2.110.0  # For Google Custom Search API
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0
aiohttp>=3.9.0
python-multipart==0.0.6
passlib==1.7.4
PyJWT[crypto]==2.8.0
bcrypt==4.0.1
httpx==0.25.0
pytest==7.4.3
pytest-asyncio==0.21.1
cachetools>=5.3.2  # For caching search results
numpy>=1.26.0  # For data processing
pandas>=2.1.0  # For data analysis
scikit-learn>=1.3.2  # For simple ML operations (optional)
reportlab>=4.0.0  # For PDF generation
Pillow>=10.0.0  # For image processing in PDFs
